@@ -1,14 +1,17 @@
 package com.listenergao.audioandvideolearning.adapter;
 
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import com.chad.library.adapter.base.BaseQuickAdapter;
-import com.chad.library.adapter.base.BaseViewHolder;
+import com.chad.library.adapter.base.viewholder.BaseViewHolder;
 import com.listenergao.audioandvideolearning.R;
 import com.listenergao.audioandvideolearning.mode.CategoryBean;
 
 import java.util.List;
 
+/**
+ * @author listenergao
+ */
 public class CategoryAdapter extends BaseQuickAdapter<CategoryBean, BaseViewHolder> {
 
     public CategoryAdapter(@Nullable List<CategoryBean> data) {
