@@ -1,15 +1,17 @@
 package com.listenergao.audioandvideolearning.activity;
 
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 
-import com.listenergao.audioandvideolearning.R;
+import androidx.appcompat.app.AppCompatActivity;
 
+
+/**
+ * @author listenergao
+ */
 public class BaseActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_base);
     }
 }
